test(*): Test new Scss setup
