+++ config/constants.js
@@ -1,0 +1,7 @@
+const input = 'src';
+const output = 'dist';
+
+module.exports = {
+  input,
+  output,
+};
