+++ src/index.njk
@@ -5,12 +5,10 @@ layout: default
 sitemap_priority: 1
 no_logo: true
 hero_h1: Frontend developer, typography enthusiast, amateur triathlete, & LEGO master builder
+custom_css: pages/home.scss
 ---
 
 {% from "components/hero.html" import hero %}
-
-{% set css %}{% include "pages/home.scss" %}{% endset %}
-<style>{{ css | scss | safe }}</style>
 
 {% set h1 %}Frontend dev<span class="home-hero-hide">eloper</span>, typo<span class="home-hero-hide">graphy</span> enthusiast, <span class="home-hero-hide">amateur </span>triathlete, &amp;&nbsp;<span class="sc">LEGO</span> master builder.{% endset %}
